+++ src/api/actions.ts
@@ -1,11 +1,11 @@
 import { Collection, GuildMember, Role } from "discord.js";
-import Main from "../../Main";
-import logger from "../../utils/logger";
-import { ManageRolesParams } from "../types/params";
-import { ActionError, UserResult } from "../types/results";
-import { getUserResult } from "../../utils/utils";
+import Main from "../Main";
+import logger from "../utils/logger";
+import { ManageRolesParams } from "./types/params";
+import { ActionError, InviteResult, UserResult } from "./types/results";
+import { getUserResult } from "../utils/utils";
 
-export default async function manageRoles(
+export async function manageRoles(
   params: ManageRolesParams,
   isUpgrade: boolean
 ): Promise<UserResult> {
@@ -36,4 +36,29 @@ export default async function manageRoles(
   updatedMember.send(params.message).catch(logger.error);
 
   return getUserResult(updatedMember);
+}
+
+export async function generateInvite(guildId: string): Promise<InviteResult> {
+  const guild = await Main.Client.guilds.fetch(guildId);
+
+  const invite = await guild.systemChannel.createInvite({
+    maxAge: 60 * 15,
+    maxUses: 1,
+    unique: true,
+  });
+
+  return {
+    code: invite.code,
+  };
+}
+
+export async function isMember(
+  guildId: string,
+  userId: string
+): Promise<UserResult> {
+  const guild = await Main.Client.guilds.fetch(guildId);
+
+  const member = await guild.members.fetch(userId);
+
+  return getUserResult(member);
 }
